+++ single_game.py
@@ -82,10 +82,10 @@ class SingleGame:
         self.card_count = 7  # 처음 시작하는 카드 수
         self.winner_message = ""  # 승리 메세지
 
-        # 로비에서 가져온 정보
         self.computer_attends = computer_attends
-        for i in range(len(self.computer_attends)):
-            if self.computer_attends[i]:
+        # 로비에서 가져온 정보
+        for attend in self.computer_attends:
+            if attend:
                 self.player_count += 1
 
         self.computer_color = ["red", "blue", "green", "yellow"]
@@ -275,7 +275,7 @@ class SingleGame:
         self.uno_remaining_time_text = None  # 유저턴 남은 우노 시간 텍스트
         self.com_uno_remaining_time = None  # 컴퓨터턴 남은 우노 시간
         self.com_uno_remaining_time_text = None  # 컴퓨터턴 남은 우노 시간 텍스트
-
+        self.turn_count = 1 # 턴 카운트
         # change카드 좌표, spacing은 공백
         self.x5 = 100
         self.y5 = 100
@@ -349,6 +349,9 @@ class SingleGame:
         self.com_uno_remaining_time_text = None  # 컴퓨터턴 남은 우노 시간 텍스트
 
     def game(self):
+        print(len(self.remain_cards))
+        print(len(self.player_hands[0]), len(self.player_hands[1]), len(self.player_hands[2]), len(self.player_hands[3]))
+        print(len(self.player_hands))
         # 마우스의 위치를 가져옴
         mouse_x, mouse_y = pygame.mouse.get_pos()
         # 현재 플레이어 결정
@@ -394,6 +397,7 @@ class SingleGame:
                 elif self.new_drawn_card is not None and self.clicked_next_turn_button:
                     self.current_player = (self.current_player + self.game_direction) % self.player_count
                     self.reset()
+                    self.turn_count = self.turn_count + 1
                 # 카드를 드로우 하고, 드로우한 카드를 내는 함수.
                 elif self.new_drawn_card is not None and self.pop_card is None:
                     # 유효성 검사 및 클릭카드가 new_drawn_card인지 확인
@@ -432,6 +436,7 @@ class SingleGame:
                                                                                          self.player_count)
                             self.animation_method[self.pop_card.value](self.current_player)
                             self.reset()
+                            self.turn_count = self.turn_count + 1
                         # 내는 카드가 special이고, change일 경우
                         elif self.pop_card.is_special() and self.pop_card.value == "change":
                             self.change_card = True
@@ -443,10 +448,12 @@ class SingleGame:
                                 self.animation_method[self.pop_card.value](self.current_player)
                                 self.board_card.append(self.color_change)
                                 self.reset()
+                                self.turn_count = self.turn_count + 1
                         # 내는 카드가 special이 아닌 경우
                         elif not self.pop_card.is_special():
                             self.current_player = (self.current_player + self.game_direction) % self.player_count
                             self.reset()
+                            self.turn_count = self.turn_count + 1
                 elif not self.uno_check:
                     # 내는 카드가 special이고 change가 아닐 때,
                     if self.pop_card.is_special() and self.pop_card.value != "change":
@@ -458,6 +465,7 @@ class SingleGame:
                                                                                      self.player_count)
                         self.animation_method[self.pop_card.value](self.current_player)
                         self.reset()
+                        self.turn_count = self.turn_count + 1
                     # 내는 카드가 change인 경우
                     elif self.pop_card.value == "change":
                         self.change_card = True
@@ -472,10 +480,12 @@ class SingleGame:
                             self.animation_method[self.pop_card.value](self.current_player)
                             self.board_card.append(self.color_change)
                             self.reset()
+                            self.turn_count = self.turn_count + 1
                     # 내는 카드가 special이 아닌 경우
                     else:
                         self.current_player = (self.current_player + self.game_direction) % self.player_count
                         self.reset()
+                        self.turn_count = self.turn_count + 1
 
             # 우노 시간 넘기면 발동
             if self.pop_card is not None and self.one_flags[0]:
@@ -497,11 +507,13 @@ class SingleGame:
                             self.player_count)
                         self.animation_method[self.pop_card.value](self.current_player)
                         self.reset()
+                        self.turn_count = self.turn_count + 1
 
                     # 내는 카드가 special이 아닌 경우
                     elif not self.pop_card.is_special():
                         self.current_player = (self.current_player + self.game_direction) % self.player_count
                         self.reset()
+                        self.turn_count = self.turn_count + 1
             # 우노 마치고, change_card 제한시간 넘는 경우
             elif self.change_uno_expire:
                 if self.user_turn and self.current_time - self.turn_start_time >= self.time_limit:
@@ -511,6 +523,7 @@ class SingleGame:
                         self.player_count)
                     self.animation_method[self.pop_card.value](self.current_player)
                     self.reset()
+                    self.turn_count = self.turn_count + 1
             # 유저의 턴일 때 시간이 초과되면 드로우하고 턴을 넘김
             elif self.new_drawn_card is None:
                 if self.user_turn and self.current_time - self.turn_start_time >= self.time_limit:
@@ -519,12 +532,14 @@ class SingleGame:
                     self.player_hands[self.current_player].append(self.new_drawn_card)  # 카드를 드로우
                     self.current_player = (self.current_player + self.game_direction) % self.player_count
                     self.reset()
+                    self.turn_count = self.turn_count + 1
 
             # 유저가 드로우를 하고 시간이 초과되면 턴을 그냥 넘김
             elif self.new_drawn_card is not None:
                 if self.user_turn and self.current_time - self.user_draw_time >= self.time_limit:
                     self.current_player = (self.current_player + self.game_direction) % self.player_count
                     self.reset()
+                    self.turn_count = self.turn_count + 1
 
         # 컴퓨터 턴 처리
         if not self.user_turn:
@@ -546,6 +561,7 @@ class SingleGame:
                     self.draw_animation(self.current_player)
                     self.new_drawn_card = self.remain_cards.pop()
                     self.player_hands[self.current_player].append(self.new_drawn_card)
+                    self.turn_count = self.turn_count + 1
                 # 드로우한 카드가 낼 수 있는 경우
                 elif self.new_drawn_card is not None and is_valid_move(self.new_drawn_card, self.top_card) and self.pop_card is None:
                     if self.current_time - self.turn_start_time >= self.delay_time2:
@@ -565,6 +581,7 @@ class SingleGame:
                         self.new_drawn_card = None
                         self.computer_action_time = pygame.time.get_ticks() + self.delay_time
                         self.turn_start_time = pygame.time.get_ticks()
+                        self.turn_count = self.turn_count + 1
                 # 컴퓨터가 우노이고, 유저가 우노를 클릭했을 경우
                 elif self.uno_check:
                     if self.user_uno_clicked:
@@ -582,6 +599,7 @@ class SingleGame:
                                                                                                   self.player_count)
                             self.animation_method[self.pop_card.value](self.current_player)
                             self.reset()
+                            self.turn_count = self.turn_count + 1
                         # 컴퓨터가 낸 카드가 change일 경우
                         elif self.pop_card.value == "change":
                             self.change_index = random.randint(0, 3)
@@ -595,10 +613,12 @@ class SingleGame:
                                                                                                   self.player_count)
                             self.animation_method[self.pop_card.value](self.current_player)
                             self.reset()
+                            self.turn_count = self.turn_count + 1
                         # 컴퓨터가 낸 카드가 special이 아닌 경우
                         elif not self.pop_card.is_special():
                             self.current_player = (self.current_player + self.game_direction) % self.player_count
                             self.reset()
+                            self.turn_count = self.turn_count + 1
                 # 컴퓨터가 우노가 아니고, 컴퓨터가 카드를 낸 경우
                 elif not self.uno_check and self.pop_card is not None:
                     # 내는 카드가 special이고 change가 아닐 경우
@@ -611,6 +631,7 @@ class SingleGame:
                                                                                      self.player_count)
                         self.animation_method[self.pop_card.value](self.current_player)
                         self.reset()
+                        self.turn_count = self.turn_count + 1
                     # 내는 카드가 special이고, change일 경우
                     elif self.pop_card.is_special() and self.pop_card.value == "change":
                         if self.current_time - self.turn_start_time >= self.delay_time3:
@@ -625,10 +646,12 @@ class SingleGame:
                             self.animation_method[self.pop_card.value](self.current_player)
                             self.board_card.append(self.color_change)
                             self.reset()
+                            self.turn_count = self.turn_count + 1
                     # 내는 카드가 special이 아닌 경우
                     elif not self.pop_card.is_special():
                         self.current_player = (self.current_player + self.game_direction) % self.player_count
                         self.reset()
+                        self.turn_count = self.turn_count + 1
 
             # 컴퓨터의 턴일 때 시간이 초과되면 우노를 넘긴다.
             if any(self.one_flags[1:]) and not self.user_turn:
@@ -641,6 +664,7 @@ class SingleGame:
                             self.player_count)
                         self.animation_method[self.pop_card.value](self.current_player)
                         self.reset()
+                        self.turn_count = self.turn_count + 1
 
                     # 내는 카드가 special이고 change인 경우
                     elif self.pop_card.is_special() and self.pop_card.value == "change":
@@ -654,10 +678,12 @@ class SingleGame:
                             self.animation_method[self.pop_card.value](self.current_player)
                             self.board_card.append(self.color_change)
                             self.reset()
+                            self.turn_count = self.turn_count + 1
                     # 내는 카드가 special이 아닌 경우
                     elif not self.pop_card.is_special():
                         self.current_player = (self.current_player + self.game_direction) % self.player_count
                         self.reset()
+                        self.turn_count = self.turn_count + 1
 
     def win(self):
         if self.game_over:
@@ -854,7 +880,7 @@ class SingleGame:
             self.screen.blit(self.turn_arrow_img, (self.computer_coordinate[3][0]-self.turn_coordinate[2], self.computer_coordinate[3][1]-self.turn_coordinate[3]))
         elif self.current_player == 5:
             self.screen.blit(self.turn_arrow_img, (self.computer_coordinate[4][0]-self.turn_coordinate[2], self.computer_coordinate[4][1]-self.turn_coordinate[3]))
-
+        print(self.remain_cards)
         # 남은 카드 더미 그리기
         self.screen.blit(self.remain_cards[0].card_img_back, (self.remain_cards_x_position, self.screen.get_rect().centery - 100))
         # 엎은 카드 그리기
@@ -946,7 +972,7 @@ class SingleGame:
             Mouse.updateMouseState()
             self.clock.tick(basic.fps)
             # 카드 섞기 발생
-            if len(self.remain_cards) < 7:
+            if self.turn_count % 10 == 0:
                 self.card_shuffle_music.set_volume(self.sound_volume * self.background_volume)
                 self.card_shuffle_music.play(1)
                 self.board_card, self.remain_cards = card_reshuffle(self.board_card, self.remain_cards)
@@ -955,421 +981,5 @@ class SingleGame:
             self.event()
 
 
-class StageC(SingleGame):
-    def __init__(self):
-        super().__init__([True, False, True, False, False])
-        self.turn_count = 1
-        self.dummy_cards = generate_c_stage_cards(self.color_weakness, self.size_change)
-        self.dummy_cards_c = generate_c_for_change_cards(self.color_weakness, self.size_change)
 
-    def game(self):
-        # 마우스의 위치를 가져옴
-        mouse_x, mouse_y = pygame.mouse.get_pos()
-        # 현재 플레이어 결정
-        self.top_card = get_top_card(self.board_card)
-        # 현재 시각 불러옴
-        self.current_time = pygame.time.get_ticks()
 
-        # 플레이어턴, 컴퓨터턴 결정
-        if self.current_player == 0:
-            self.user_turn = True
-        else:
-            self.user_turn = False
-
-        if Mouse.getMouseState() == MouseState.CLICK:
-            if self.uno_button_rect.collidepoint(mouse_x, mouse_y) and len(self.player_hands[self.current_player]) == 1:
-                self.user_uno_clicked = True
-        self.hovered_card_index = find_hovered_card(self.player_hands[0], self.user_coordinate[0],
-                                                    self.user_coordinate[1], self.user_spacing, mouse_x, mouse_y, self.max_per_row)
-        self.hovered_change_index = find_hovered_change(self.change_color_list, self.x5, self.y5, self.spacing5, mouse_x, mouse_y)
-
-        # 현재 플레이어가 유저인지 확인
-        if self.user_turn:
-            if self.turn_start_time is None:
-                self.turn_start_time = pygame.time.get_ticks()
-            if Mouse.getMouseState() == MouseState.CLICK:
-                self.clicked_card_index, self.clicked_card = get_clicked_card(self.player_hands[0],
-                                                                              self.user_coordinate[0], self.user_coordinate[1], self.user_spacing, mouse_x, mouse_y, self.max_per_row)
-                # remain_cards 클릭 확인
-                self.clicked_remain_cards = self.remain_cards_rect.collidepoint(mouse_x, mouse_y)
-                # next_turn_button 클릭 확인
-                self.clicked_next_turn_button = self.next_turn_button_rect.collidepoint(mouse_x, mouse_y)
-                # change 클릭 확인
-                self.clicked_change_index, self.clicked_change = get_clicked_change(self.change_color_list, self.x5, self.y5, self.spacing5, mouse_x,
-                                                                          mouse_y)
-            if self.clicked_card is not None or self.clicked_remain_cards or self.clicked_next_turn_button or self.user_uno_clicked:
-                # 카드를 드로우함
-                if self.clicked_remain_cards and self.new_drawn_card is None and self.pop_card is None:
-                    self.draw_animation(self.current_player)
-                    self.user_draw_time = pygame.time.get_ticks()
-                    self.player_hands[0].append(self.remain_cards.pop())
-                    self.new_drawn_card = self.player_hands[0][-1]
-                # 카드를 드로우 하고 턴을 넘기는 함수.
-                elif self.new_drawn_card is not None and self.clicked_next_turn_button:
-                    self.current_player = (self.current_player + self.game_direction) % self.player_count
-                    self.reset()
-                    print("턴 종료")
-                    self.turn_count = self.turn_count + 1
-                    if self.turn_count % 5 == 0:
-                        self.board_card = random_top_card_color(self.top_card, self.dummy_cards, self.board_card, self.dummy_cards_c)
-                # 카드를 드로우 하고, 드로우한 카드를 내는 함수.
-                elif self.new_drawn_card is not None and self.pop_card is None:
-                    # 유효성 검사 및 클릭카드가 new_drawn_card인지 확인
-                    if self.new_drawn_card is not None and is_valid_move(self.new_drawn_card, self.top_card) and self.clicked_card == self.new_drawn_card and self.pop_card is None:
-                        self.place_animation(self.current_player)
-                        self.board_card, self.player_hands[self.current_player], self.pop_card = user_submit_card\
-                            (self.clicked_card, self.clicked_card_index, self.board_card, self.player_hands[self.current_player])
-                        self.uno_check = check_uno(self.player_hands[0])
-                        if self.uno_check:
-                            self.one_flags, self.uno_current_time, self.uno_delay_time = is_uno(self.one_flags,
-                                                                                                self.current_player)
-                # 카드를 드로우하지 않고, 카드를 냄
-                elif self.new_drawn_card is None and self.pop_card is None:
-                    # 유효성 검사
-                    if self.clicked_card is not None and is_valid_move(self.clicked_card, self.top_card):
-                        self.place_animation(self.current_player)
-                        self.board_card, self.player_hands[self.current_player], self.pop_card = user_submit_card\
-                            (self.clicked_card, self.clicked_card_index, self.board_card, self.player_hands[self.current_player])
-                        self.uno_check = check_uno(self.player_hands[0])
-                        if self.uno_check:
-                            self.one_flags, self.uno_current_time, self.uno_delay_time = is_uno(self.one_flags,
-                                                                                                self.current_player)
-            # 카드를 냈을 때,
-            if self.pop_card is not None:
-                # 우노일 경우
-                if self.uno_check:
-                    if self.user_uno_clicked:
-                        self.one_flags[0] = False
-                        # 내는 카드가 special이고 change가 아닐 경우
-                        if self.pop_card.is_special() and self.pop_card.value != "change":
-                            self.current_player, self.game_direction = apply_special_card_effects(self.pop_card,
-                                                                                         self.current_player,
-                                                                                         self.game_direction,
-                                                                                         self.player_hands,
-                                                                                         self.remain_cards,
-                                                                                         self.player_count)
-                            self.animation_method[self.pop_card.value](self.current_player)
-                            self.reset()
-                            print("턴 종료")
-                            self.turn_count = self.turn_count + 1
-                            if self.turn_count % 5 == 0:
-                                self.board_card = random_top_card_color(self.top_card, self.dummy_cards,
-                                                                        self.board_card, self.dummy_cards_c)
-                        # 내는 카드가 special이고, change일 경우
-                        elif self.pop_card.is_special() and self.pop_card.value == "change":
-                            self.change_card = True
-                            if self.clicked_change_index is not None:
-                                self.color_change = self.change_color_list[self.clicked_change_index]
-                                self.current_player, self.game_direction = apply_special_card_effects(
-                                    self.pop_card, self.current_player, self.game_direction,
-                                    self.player_hands, self.remain_cards, self.player_count)
-                                self.animation_method[self.pop_card.value](self.current_player)
-                                self.board_card.append(self.color_change)
-                                self.reset()
-                                print("턴 종료")
-                                self.turn_count = self.turn_count + 1
-                                if self.turn_count % 5 == 0:
-                                    self.board_card = random_top_card_color(self.top_card, self.dummy_cards,
-                                                                            self.board_card, self.dummy_cards_c)
-                        # 내는 카드가 special이 아닌 경우
-                        elif not self.pop_card.is_special():
-                            self.current_player = (self.current_player + self.game_direction) % self.player_count
-                            self.reset()
-                            print("턴 종료")
-                            self.turn_count = self.turn_count + 1
-                            if self.turn_count % 5 == 0:
-                                self.board_card = random_top_card_color(self.top_card, self.dummy_cards,
-                                                                        self.board_card, self.dummy_cards_c)
-                elif not self.uno_check:
-                    # 내는 카드가 special이고 change가 아닐 때,
-                    if self.pop_card.is_special() and self.pop_card.value != "change":
-                        self.current_player, self.game_direction = apply_special_card_effects(self.pop_card,
-                                                                                     self.current_player,
-                                                                                     self.game_direction,
-                                                                                     self.player_hands,
-                                                                                     self.remain_cards,
-                                                                                     self.player_count)
-                        self.animation_method[self.pop_card.value](self.current_player)
-                        self.reset()
-                        print("턴 종료")
-                        self.turn_count = self.turn_count + 1
-                        if self.turn_count % 5 == 0:
-                            self.board_card = random_top_card_color(self.top_card, self.dummy_cards, self.board_card, self.dummy_cards_c)
-                    # 내는 카드가 change인 경우
-                    elif self.pop_card.value == "change":
-                        self.change_card = True
-                        if self.clicked_change_index is not None:
-                            self.color_change = self.change_color_list[self.clicked_change_index]
-                            self.current_player, self.game_direction = apply_special_card_effects(self.pop_card,
-                                                                                         self.current_player,
-                                                                                         self.game_direction,
-                                                                                         self.player_hands,
-                                                                                         self.remain_cards,
-                                                                                         self.player_count)
-                            self.animation_method[self.pop_card.value](self.current_player)
-                            self.board_card.append(self.color_change)
-                            self.reset()
-                            print("턴 종료")
-                            self.turn_count = self.turn_count + 1
-                            if self.turn_count % 5 == 0:
-                                self.board_card = random_top_card_color(self.top_card, self.dummy_cards,
-                                                                        self.board_card, self.dummy_cards_c)
-                    # 내는 카드가 special이 아닌 경우
-                    else:
-                        self.current_player = (self.current_player + self.game_direction) % self.player_count
-                        self.reset()
-                        print("턴 종료")
-                        self.turn_count = self.turn_count + 1
-                        if self.turn_count % 5 == 0:
-                            self.board_card = random_top_card_color(self.top_card, self.dummy_cards, self.board_card, self.dummy_cards_c)
-
-            # 우노 시간 넘기면 발동
-            if self.pop_card is not None and self.one_flags[0]:
-                if self.current_time - self.uno_current_time >= self.uno_delay_time:
-                    self.draw_animation(self.current_player)
-                    self.uno_drawn_card = self.remain_cards.pop()
-                    self.player_hands[0].append(self.uno_drawn_card)
-                    # 내는 카드가 special이고, change일 경우
-                    if self.pop_card.is_special() and self.pop_card.value == "change":
-                        self.one_flags[0] = False
-                        self.uno_check = False
-                        self.user_uno_clicked = False
-                        self.change_uno_expire = True
-                        print("턴 종료")
-                        self.turn_count = self.turn_count + 1
-                        if self.turn_count % 5 == 0:
-                            self.board_card = random_top_card_color(self.top_card, self.dummy_cards, self.board_card, self.dummy_cards_c)
-                    # 내는 카드가 special이고, change가 아닐 경우
-                    elif self.pop_card.is_special() and self.pop_card.value != "change":
-                        self.current_player, self.game_direction = apply_special_card_effects(
-                            self.pop_card, self.current_player, self.game_direction, self.player_hands,
-                            self.remain_cards,
-                            self.player_count)
-                        self.animation_method[self.pop_card.value](self.current_player)
-                        self.reset()
-                        print("턴 종료")
-                        self.turn_count = self.turn_count + 1
-                        if self.turn_count % 5 == 0:
-                            self.board_card = random_top_card_color(self.top_card, self.dummy_cards, self.board_card, self.dummy_cards_c)
-
-                    # 내는 카드가 special이 아닌 경우
-                    elif not self.pop_card.is_special():
-                        self.current_player = (self.current_player + self.game_direction) % self.player_count
-                        self.reset()
-                        print("턴 종료")
-                        self.turn_count = self.turn_count + 1
-                        if self.turn_count % 5 == 0:
-                            self.board_card = random_top_card_color(self.top_card, self.dummy_cards, self.board_card, self.dummy_cards_c)
-            # 우노 마치고, change_card 제한시간 넘는 경우
-            elif self.change_uno_expire:
-                if self.user_turn and self.current_time - self.turn_start_time >= self.time_limit:
-                    self.current_player, self.game_direction = apply_special_card_effects(
-                        self.pop_card, self.current_player, self.game_direction, self.player_hands,
-                        self.remain_cards,
-                        self.player_count)
-                    self.animation_method[self.pop_card.value](self.current_player)
-                    self.reset()
-                    print("턴 종료")
-                    self.turn_count = self.turn_count + 1
-                    if self.turn_count % 5 == 0:
-                        self.board_card = random_top_card_color(self.top_card, self.dummy_cards, self.board_card, self.dummy_cards_c)
-            # 유저의 턴일 때 시간이 초과되면 드로우하고 턴을 넘김
-            elif self.new_drawn_card is None:
-                if self.user_turn and self.current_time - self.turn_start_time >= self.time_limit:
-                    self.draw_animation(self.current_player)
-                    self.new_drawn_card = self.remain_cards.pop()
-                    self.player_hands[self.current_player].append(self.new_drawn_card)  # 카드를 드로우
-                    self.current_player = (self.current_player + self.game_direction) % self.player_count
-                    self.reset()
-                    print("턴 종료")
-                    self.turn_count = self.turn_count + 1
-                    if self.turn_count % 5 == 0:
-                        self.board_card = random_top_card_color(self.top_card, self.dummy_cards, self.board_card, self.dummy_cards_c)
-
-            # 유저가 드로우를 하고 시간이 초과되면 턴을 그냥 넘김
-            elif self.new_drawn_card is not None:
-                if self.user_turn and self.current_time - self.user_draw_time >= self.time_limit:
-                    self.current_player = (self.current_player + self.game_direction) % self.player_count
-                    self.reset()
-                    print("턴 종료")
-                    self.turn_count = self.turn_count + 1
-                    if self.turn_count % 5 == 0:
-                        self.board_card = random_top_card_color(self.top_card, self.dummy_cards, self.board_card, self.dummy_cards_c)
-
-        # 컴퓨터 턴 처리
-        if not self.user_turn:
-            if self.current_time >= self.computer_action_time:  # 설정한 시간이 되면 컴퓨터가 행동함
-                # 처음 유효성 검사
-                if self.new_drawn_card is None and self.pop_card is None:
-                    self.playable, self.pop_card_index = computer_playable_card(self.player_hands[self.current_player], self.board_card)
-                # 카드를 낼 수 있을 때 낸다.
-                if self.playable and self.new_drawn_card is None and self.pop_card is None:
-                    self.place_animation(self.current_player)
-                    self.pop_card = self.player_hands[self.current_player][self.pop_card_index]
-                    self.board_card, self.player_hands[self.current_player] = com_submit_card\
-                        (self.pop_card, self.pop_card_index, self.board_card, self.player_hands[self.current_player])
-                    self.uno_check = check_uno(self.player_hands[self.current_player])
-                    if self.uno_check:
-                        self.one_flags, self.uno_current_time, self.uno_delay_time = is_uno(self.one_flags, self.current_player)
-                # 카드를 낼 수 없을 때 드로우 한다.
-                elif not self.playable and self.new_drawn_card is None and self.pop_card is None:
-                    self.draw_animation(self.current_player)
-                    self.new_drawn_card = self.remain_cards.pop()
-                    self.player_hands[self.current_player].append(self.new_drawn_card)
-                # 드로우한 카드가 낼 수 있는 경우
-                elif self.new_drawn_card is not None and is_valid_move(self.new_drawn_card, self.top_card) and self.pop_card is None:
-                    if self.current_time - self.turn_start_time >= self.delay_time2:
-                        self.place_animation(self.current_player)
-                        self.pop_card = self.new_drawn_card
-                        self.pop_card_index = self.player_hands[self.current_player].index(self.pop_card)
-                        self.board_card, self.player_hands[self.current_player] = com_submit_card \
-                            (self.pop_card, self.pop_card_index, self.board_card, self.player_hands[self.current_player])
-                        self.uno_check = check_uno(self.player_hands[self.current_player])
-                        if self.uno_check:
-                            self.one_flags, self.uno_current_time, self.uno_delay_time = is_uno(self.one_flags,
-                                                                                                self.current_player)
-                # 드로우한 카드를 낼 수 없는 경우
-                elif self.new_drawn_card is not None and not is_valid_move(self.new_drawn_card, self.top_card) and self.pop_card is None:
-                    if self.current_time - self.turn_start_time >= self.delay_time2:
-                        self.current_player = (self.current_player + self.game_direction) % self.player_count
-                        self.new_drawn_card = None
-                        self.computer_action_time = pygame.time.get_ticks() + self.delay_time
-                        self.turn_start_time = pygame.time.get_ticks()
-                        print("턴 종료")
-                        self.turn_count = self.turn_count + 1
-                        if self.turn_count % 5 == 0:
-                            self.board_card = random_top_card_color(self.top_card, self.dummy_cards, self.board_card, self.dummy_cards_c)
-                # 컴퓨터가 우노이고, 유저가 우노를 클릭했을 경우
-                elif self.uno_check:
-                    if self.user_uno_clicked:
-                        self.draw_animation(self.current_player)
-                        self.one_flags[self.current_player] = False
-                        self.uno_drawn_card = self.remain_cards.pop()
-                        self.player_hands[self.current_player].append(self.uno_drawn_card)
-                        # 컴퓨터가 낸 카드가 special일 경우
-                        if self.pop_card.is_special() and self.pop_card.value != "change":
-                            self.current_player, self.game_direction = apply_special_card_effects(self.pop_card,
-                                                                                                  self.current_player,
-                                                                                                  self.game_direction,
-                                                                                                  self.player_hands,
-                                                                                                  self.remain_cards,
-                                                                                                  self.player_count)
-                            self.animation_method[self.pop_card.value](self.current_player)
-                            self.reset()
-                            print("턴 종료")
-                            self.turn_count = self.turn_count + 1
-                            if self.turn_count % 5 == 0:
-                                self.board_card = random_top_card_color(self.top_card, self.dummy_cards,
-                                                                        self.board_card, self.dummy_cards_c)
-                        # 컴퓨터가 낸 카드가 change일 경우
-                        elif self.pop_card.value == "change":
-                            self.change_index = random.randint(0, 3)
-                            self.color_change = self.change_color_list[self.change_index]
-                            self.board_card.append(self.color_change)
-                            self.current_player, self.game_direction = apply_special_card_effects(self.pop_card,
-                                                                                                  self.current_player,
-                                                                                                  self.game_direction,
-                                                                                                  self.player_hands,
-                                                                                                  self.remain_cards,
-                                                                                                  self.player_count)
-                            self.animation_method[self.pop_card.value](self.current_player)
-                            self.reset()
-                            print("턴 종료")
-                            self.turn_count = self.turn_count + 1
-                            if self.turn_count % 5 == 0:
-                                self.board_card = random_top_card_color(self.top_card, self.dummy_cards,
-                                                                        self.board_card, self.dummy_cards_c)
-                        # 컴퓨터가 낸 카드가 special이 아닌 경우
-                        elif not self.pop_card.is_special():
-                            self.current_player = (self.current_player + self.game_direction) % self.player_count
-                            self.reset()
-                            print("턴 종료")
-                            self.turn_count = self.turn_count + 1
-                            if self.turn_count % 5 == 0:
-                                self.board_card = random_top_card_color(self.top_card, self.dummy_cards,
-                                                                        self.board_card, self.dummy_cards_c)
-                # 컴퓨터가 우노가 아니고, 컴퓨터가 카드를 낸 경우
-                elif not self.uno_check and self.pop_card is not None:
-                    # 내는 카드가 special이고 change가 아닐 경우
-                    if self.pop_card.is_special() and self.pop_card.value != "change":
-                        self.current_player, self.game_direction = apply_special_card_effects(self.pop_card,
-                                                                                     self.current_player,
-                                                                                     self.game_direction,
-                                                                                     self.player_hands,
-                                                                                     self.remain_cards,
-                                                                                     self.player_count)
-                        self.animation_method[self.pop_card.value](self.current_player)
-                        self.reset()
-                        print("턴 종료")
-                        self.turn_count = self.turn_count + 1
-                        if self.turn_count % 5 == 0:
-                            self.board_card = random_top_card_color(self.top_card, self.dummy_cards, self.board_card, self.dummy_cards_c)
-                    # 내는 카드가 special이고, change일 경우
-                    elif self.pop_card.is_special() and self.pop_card.value == "change":
-                        if self.current_time - self.turn_start_time >= self.delay_time3:
-                            self.change_index = random.randint(0, 3)
-                            self.color_change = self.change_color_list[self.change_index]
-                            self.current_player, self.game_direction = apply_special_card_effects(self.pop_card,
-                                                                                         self.current_player,
-                                                                                         self.game_direction,
-                                                                                         self.player_hands,
-                                                                                         self.remain_cards,
-                                                                                         self.player_count)
-                            self.animation_method[self.pop_card.value](self.current_player)
-                            self.board_card.append(self.color_change)
-                            self.reset()
-                            print("턴 종료")
-                            self.turn_count = self.turn_count + 1
-                            if self.turn_count % 5 == 0:
-                                self.board_card = random_top_card_color(self.top_card, self.dummy_cards,
-                                                                        self.board_card, self.dummy_cards_c)
-                    # 내는 카드가 special이 아닌 경우
-                    elif not self.pop_card.is_special():
-                        self.current_player = (self.current_player + self.game_direction) % self.player_count
-                        self.reset()
-                        print("턴 종료")
-                        self.turn_count = self.turn_count + 1
-                        if self.turn_count % 5 == 0:
-                            self.board_card = random_top_card_color(self.top_card, self.dummy_cards, self.board_card, self.dummy_cards_c)
-
-            # 컴퓨터의 턴일 때 시간이 초과되면 우노를 넘긴다.
-            if any(self.one_flags[1:]) and not self.user_turn:
-                if self.current_time - self.uno_current_time >= self.uno_delay_time:
-                    # 내는 카드가 special이고, change가 아닐 경우
-                    if self.pop_card.is_special() and self.pop_card.value != "change":
-                        self.current_player, self.game_direction = apply_special_card_effects(
-                            self.pop_card, self.current_player, self.game_direction, self.player_hands,
-                            self.remain_cards,
-                            self.player_count)
-                        self.animation_method[self.pop_card.value](self.current_player)
-                        self.reset()
-                        print("턴 종료")
-                        self.turn_count = self.turn_count + 1
-                        if self.turn_count % 5 == 0:
-                            self.board_card = random_top_card_color(self.top_card, self.dummy_cards, self.board_card, self.dummy_cards_c)
-
-                    # 내는 카드가 special이고 change인 경우
-                    elif self.pop_card.is_special() and self.pop_card.value == "change":
-                        self.change_index = random.randint(0, 3)
-                        self.color_change = self.change_color_list[self.change_index]
-                        if self.current_time - self.turn_start_time >= self.delay_time3:
-                            self.current_player, self.game_direction = apply_special_card_effects(
-                                self.pop_card, self.current_player, self.game_direction, self.player_hands,
-                                self.remain_cards,
-                                self.player_count)
-                            self.animation_method[self.pop_card.value](self.current_player)
-                            self.board_card.append(self.color_change)
-                            self.reset()
-                            print("턴 종료")
-                            self.turn_count = self.turn_count + 1
-                            if self.turn_count % 5 == 0:
-                                self.board_card = random_top_card_color(self.top_card, self.dummy_cards,
-                                                                        self.board_card, self.dummy_cards_c)
-                    # 내는 카드가 special이 아닌 경우
-                    elif not self.pop_card.is_special():
-                        self.current_player = (self.current_player + self.game_direction) % self.player_count
-                        self.reset()
-                        print("턴 종료")
-                        self.turn_count = self.turn_count + 1
-                        if self.turn_count%5 == 0:
-                            self.board_card = random_top_card_color(self.top_card, self.dummy_cards, self.board_card, self.dummy_cards_c)
