+++ single_play_lobby.py
@@ -4,7 +4,7 @@ import random
 import math
 
 from single_game import SingleGame
-from single_game import StageC
+from stage_logic import StageB, StageC
 from mouse import Mouse, MouseState
 from slider import Slider
 from button import Button
